Extract Emoji type to its own file
