+++ app/src/lib/stores/app-store.ts
@@ -217,7 +217,8 @@ import { MergeTreeResult } from '../../models/merge'
 import { promiseWithMinimumTimeout } from '../promise'
 import { BackgroundFetcher } from './helpers/background-fetcher'
 import { RepositoryStateCache } from './repository-state-cache'
-import { Emoji, readEmoji } from '../read-emoji'
+import { readEmoji } from '../read-emoji'
+import { Emoji } from '../emoji'
 import { GitStoreCache } from './git-store-cache'
 import { GitErrorContext } from '../git-error-context'
 import {
@@ -1463,8 +1464,8 @@ export class AppStore extends TypedBaseStore<IAppState> {
     // and it also exists in the repository
     const defaultBranch =
       currentBranch != null &&
-      cachedDefaultBranch != null &&
-      currentBranch.name !== cachedDefaultBranch.name
+        cachedDefaultBranch != null &&
+        currentBranch.name !== cachedDefaultBranch.name
         ? cachedDefaultBranch
         : null
 
@@ -1702,9 +1703,9 @@ export class AppStore extends TypedBaseStore<IAppState> {
     const changesetData = await gitStore.performFailableOperation(() =>
       currentSHAs.length > 1
         ? getCommitRangeChangedFiles(
-            repository,
-            this.orderShasByHistory(repository, currentSHAs)
-          )
+          repository,
+          this.orderShasByHistory(repository, currentSHAs)
+        )
         : getChangedFiles(repository, currentSHAs[0])
     )
     if (!changesetData) {
@@ -1781,17 +1782,17 @@ export class AppStore extends TypedBaseStore<IAppState> {
     const diff =
       shas.length > 1
         ? await getCommitRangeDiff(
-            repository,
-            file,
-            this.orderShasByHistory(repository, shas),
-            this.hideWhitespaceInHistoryDiff
-          )
+          repository,
+          file,
+          this.orderShasByHistory(repository, shas),
+          this.hideWhitespaceInHistoryDiff
+        )
         : await getCommitDiff(
-            repository,
-            file,
-            shas[0],
-            this.hideWhitespaceInHistoryDiff
-          )
+          repository,
+          file,
+          shas[0],
+          this.hideWhitespaceInHistoryDiff
+        )
 
     const stateAfterLoad = this.repositoryStateCache.get(repository)
     const { shas: shasAfter } = stateAfterLoad.commitSelection
@@ -2748,7 +2749,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
       if (
         multiCommitOperationState !== null &&
         multiCommitOperationState.operationDetail.kind ===
-          MultiCommitOperationKind.CherryPick &&
+        MultiCommitOperationKind.CherryPick &&
         multiCommitOperationState.operationDetail.sourceBranch !== null
       ) {
         theirBranch =
@@ -2786,7 +2787,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
     if (
       multiCommitOperationState !== null &&
       multiCommitOperationState.operationDetail.kind ===
-        MultiCommitOperationKind.Merge &&
+      MultiCommitOperationKind.Merge &&
       multiCommitOperationState.operationDetail.sourceBranch !== null
     ) {
       theirBranch = multiCommitOperationState.operationDetail.sourceBranch.name
@@ -3056,7 +3057,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
 
       const currentFiles =
         stashEntry !== null &&
-        stashEntry.files.kind === StashedChangesLoadStates.Loaded
+          stashEntry.files.kind === StashedChangesLoadStates.Loaded
           ? stashEntry.files.files
           : []
 
@@ -3151,7 +3152,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
     if (
       changesStateAfterLoad.selection.kind !== ChangesSelectionKind.Stash ||
       changesStateAfterLoad.selection.selectedStashedFile !==
-        selectionBeforeLoad.selectedStashedFile
+      selectionBeforeLoad.selectedStashedFile
     ) {
       return
     }
@@ -6104,11 +6105,10 @@ export class AppStore extends TypedBaseStore<IAppState> {
     return `The following paths aren't Git repositories:\n\n${invalidPaths
       .slice(0, MaxInvalidFoldersToDisplay)
       .map(path => `- ${path}`)
-      .join('\n')}${
-      invalidPaths.length > MaxInvalidFoldersToDisplay
+      .join('\n')}${invalidPaths.length > MaxInvalidFoldersToDisplay
         ? `\n\n(and ${invalidPaths.length - MaxInvalidFoldersToDisplay} more)`
         : ''
-    }`
+      }`
   }
 
   private async withRefreshedGitHubRepository<T>(
@@ -6382,8 +6382,8 @@ export class AppStore extends TypedBaseStore<IAppState> {
     const encodedBaseBranch =
       baseBranch !== undefined
         ? baseForkPreface +
-          encodeURIComponent(baseBranch.nameWithoutRemote) +
-          '...'
+        encodeURIComponent(baseBranch.nameWithoutRemote) +
+        '...'
         : ''
 
     const compareForkPreface = isForkContributingToParent
@@ -6523,8 +6523,8 @@ export class AppStore extends TypedBaseStore<IAppState> {
       this.emitError(
         new Error(
           `Couldn't find branch '${headRefName}' in remote '${remote.name}'. ` +
-            `A common reason for this is that the PR author has deleted their ` +
-            `branch or their forked repository.`
+          `A common reason for this is that the PR author has deleted their ` +
+          `branch or their forked repository.`
         )
       )
       return
@@ -6671,7 +6671,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
       changesState.conflictState === null ||
       multiCommitOperationState === null ||
       multiCommitOperationState.step.kind !==
-        MultiCommitOperationStepKind.ShowConflicts
+      MultiCommitOperationStepKind.ShowConflicts
     ) {
       return
     }
@@ -7694,13 +7694,13 @@ export class AppStore extends TypedBaseStore<IAppState> {
     const changesetData =
       commitsBetweenBranches.length > 0
         ? await gitStore.performFailableOperation(() =>
-            getBranchMergeBaseChangedFiles(
-              repository,
-              baseBranch.name,
-              currentBranch.name,
-              commitsBetweenBranches[0]
-            )
+          getBranchMergeBaseChangedFiles(
+            repository,
+            baseBranch.name,
+            currentBranch.name,
+            commitsBetweenBranches[0]
           )
+        )
         : emptyChangeSet
 
     if (changesetData === undefined) {
@@ -7726,10 +7726,10 @@ export class AppStore extends TypedBaseStore<IAppState> {
       mergeStatus:
         commitSHAs.length > 0 || !hasMergeBase
           ? {
-              kind: hasMergeBase
-                ? ComputedAction.Loading
-                : ComputedAction.Invalid,
-            }
+            kind: hasMergeBase
+              ? ComputedAction.Loading
+              : ComputedAction.Invalid,
+          }
           : null,
     })
