+++ app/src/lib/text-token-parser.ts
@@ -5,7 +5,7 @@ import {
 } from '../models/repository'
 import { GitHubRepository } from '../models/github-repository'
 import { getHTMLURL } from './api'
-import { Emoji } from './read-emoji'
+import { Emoji } from './emoji'
 
 export enum TokenType {
   /*
