+++ app/src/ui/changes/undo-commit.tsx
@@ -4,7 +4,7 @@ import { Commit } from '../../models/commit'
 import { RichText } from '../lib/rich-text'
 import { RelativeTime } from '../relative-time'
 import { Button } from '../lib/button'
-import { Emoji } from '../../lib/read-emoji'
+import { Emoji } from '../../lib/emoji'
 
 interface IUndoCommitProps {
   /** The function to call when the Undo button is clicked. */
