+++ app/src/ui/drag-elements/commit-drag-element.tsx
@@ -10,7 +10,7 @@ import { CommitListItem } from '../history/commit-list-item'
 import { Octicon } from '../octicons'
 import * as octicons from '../octicons/octicons.generated'
 import { Account } from '../../models/account'
-import { Emoji } from '../../lib/read-emoji'
+import { Emoji } from '../../lib/emoji'
 
 interface ICommitDragElementProps {
   readonly commit: Commit
