+++ app/src/ui/history/unreachable-commits-dialog.tsx
@@ -6,7 +6,7 @@ import { Commit } from '../../models/commit'
 import { CommitList } from './commit-list'
 import { LinkButton } from '../lib/link-button'
 import { Account } from '../../models/account'
-import { Emoji } from '../../lib/read-emoji'
+import { Emoji } from '../../lib/emoji'
 
 export enum UnreachableCommitsTab {
   Unreachable,
