+++ app/src/ui/multi-commit-operation/base-multi-commit-operation.tsx
@@ -14,7 +14,7 @@ import { WarnForcePushDialog } from './dialog/warn-force-push-dialog'
 import { PopupType } from '../../models/popup'
 import { Account } from '../../models/account'
 import { IAPIRepoRuleset } from '../../lib/api'
-import { Emoji } from '../../lib/read-emoji'
+import { Emoji } from '../../lib/emoji'
 
 export interface IMultiCommitOperationProps {
   readonly repository: Repository
