+++ app/src/ui/multi-commit-operation/dialog/progress-dialog.tsx
@@ -5,7 +5,7 @@ import { Dialog, DialogContent } from '../../dialog'
 import { Octicon } from '../../octicons'
 import * as octicons from '../../octicons/octicons.generated'
 import { IMultiCommitOperationProgress } from '../../../models/progress'
-import { Emoji } from '../../../lib/read-emoji'
+import { Emoji } from '../../../lib/emoji'
 
 interface IProgressDialogProps {
   /**
